Fix: Remove default wildcard on exclude
Part of  [story #18320](https://tuleap.net/plugins/tracker/?aid=18320)
native support of pull requests in tuleap branch source jenkins plugin

When a new Tuleap Project is created, the default filter includes all
repositories instead of exclude all of them.

How to test:
 - Create a new Tuleap Project.
=> In the Configuration menu => Behaviors => Filter by name, here is no "*" anymore for "Exclude" field
