+++ src/test/resources/org/jenkinsci/plugins/tuleap_git_branch_source/TuleapSCMNavigatorTest/new_project_by_default.xml
@@ -1,12 +1,12 @@
-<org.jenkinsci.plugins.tuleap__git__branch__source.TuleapSCMNavigator plugin="tuleap-git-branch-source@1.0.0-SNAPSHOT">
+<org.jenkinsci.plugins.tuleap__git__branch__source.TuleapSCMNavigator>
 	<projectId>3280</projectId>
 	<traits>
 		<org.jenkinsci.plugins.tuleap__git__branch__source.trait.UserForkRepositoryTrait>
 			<strategy>1</strategy>
 		</org.jenkinsci.plugins.tuleap__git__branch__source.trait.UserForkRepositoryTrait>
-		<jenkins.scm.impl.trait.WildcardSCMSourceFilterTrait plugin="scm-api@2.2.3">
+		<jenkins.scm.impl.trait.WildcardSCMSourceFilterTrait>
 			<includes>*</includes>
-			<excludes>*</excludes>
+			<excludes/>
 		</jenkins.scm.impl.trait.WildcardSCMSourceFilterTrait>
 	</traits>
 	<credentialsId>fe09fd0e-7287-44a1-b0b5-746accd227c1</credentialsId>
